modules/gcs2spanner/src/requirements.txt
cloudevents>=1.10.0
functions-framework>=3.8.0
google-auth==2.11.0
google-cloud-logging==3.2.5
pytest==8.1.1
requests==2.20.0
